fix: 컴포넌트 재사용
